+++ src/core/services/client.js
@@ -1,7 +1,7 @@
-import { name, version } from '../../../package.json';
-
 export default class Client {
+
   constructor() {
-    this.defaultHeaders = { 'user-agent': `${name}/${version}` };
+    this.defaultHeaders = {};
   }
+
 }